V- 1.0.0

+++ cliente/listar-cliente.php
@@ -9,6 +9,7 @@
     if($qtd > 0){
         print "<table class='table table-striped table-bordered table-hover overflow-x-auto '>";
         print "<tr>";
+            print "<th>Logo</th>";
             print "<th>Empresa</th>";
             print "<th>Fundação da Empresa</th>";
             print "<th>Telefone Empresa</th>";
@@ -24,22 +25,25 @@
             
             $data_aniversario = $row->data_aniversario_contato;
             $dataFormatadaAniversario = date('d/m/y', strtotime($data_aniversario));
-            print "<tr>";
-            print "<td>".$row->nome_empresa."</td>";
-            print "<td>".$dataFormatadaEmpresa."</td>";
-            print "<td>".$row->Telefone_empresa."</td>";
-            print "<td>".$row->Nome_contato."</td>";
-            print "<td>".$dataFormatadaAniversario."</td>";
-            print "<td>".$row->email."</td>";
-            print "<td>".$row->Telefone_Contador."</td>";
-            print "<td><button onclick=\"location.href='?page_cliente=editar_cliente&id=".$row->id."'\" class='btn btn-success'>
-                            Editar
-                        </button>
-                        <button onclick=\"if(confirm('Tem certeza que deseja excluir?')){location.href='?page_cliente=salvar_cliente&acao=excluir_cliente&id=".$row->id."';}else{false;}\" class='btn btn-danger'>
-                            Excluir
-                        </button>
-                    </td>";
-            print "<tr>";
+            if($_SESSION['id'] === $row->id_contador && password_verify($_SESSION['senha'], $row->senha)){
+                print "<tr>";
+                print "<td><img src='$row->Logo.' class='img-fluid' style='width: 50px'></td>";
+                print "<td>".$row->nome_empresa."</td>";
+                print "<td>".$dataFormatadaEmpresa."</td>";
+                print "<td>".$row->Telefone_empresa."</td>";
+                print "<td>".$row->Nome_contato."</td>";
+                print "<td>".$dataFormatadaAniversario."</td>";
+                print "<td>".$row->email."</td>";
+                print "<td>".$row->Telefone_Contador."</td>";
+                print "<td><button onclick=\"location.href='?page_cliente=editar_cliente&id=".$row->id."'\" class='btn btn-success'>
+                                Editar
+                            </button>
+                            <button onclick=\"if(confirm('Tem certeza que deseja excluir?')){location.href='?page_cliente=salvar_cliente&acao=excluir_cliente&id=".$row->id."';}else{false;}\" class='btn btn-danger'>
+                                Excluir
+                            </button>
+                        </td>";
+                print "<tr>";
+            }
         }
         print "</table>";
     }else{
